Refactor auto-pagination with reusable functions
